requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrdict==2.0.0
attrs==19.2.0
awscli==1.16.111
backcall==0.1.0
base58==1.0.3
bleach==3.0.2
botocore==1.12.101
BTrees==4.4.1
cachetools==2.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
coincurve==9.0.0
colander==1.7.0
colorama==0.3.9
colorlover==0.3.0
cycler==0.10.0
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
ecdsa==0.13
entrypoints==0.2.3
eth-abi==1.2.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==2.0.0
eth-utils==1.3.0
ethereum==2.3.2
ethereum-input-decoder==0.2.2
ez-setup==0.9
Flask==0.12.2
functional==0.4
future==0.17.1
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-bigquery==1.5.1
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.7.1
hexbytes==0.1.0
idna==2.6
ijson==2.3
importlib-metadata==0.23
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.9.6
jmespath==0.9.3
json-rpc==1.11.1
jsonmerge==1.5.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
libeospy==1.1.10
lru-dict==1.1.6
MarkupSafe==1.0
matplotlib==2.2.0
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.0
numpy==1.15.4
oauthlib==2.0.4
packaging==19.2
pandas==0.23.4
pandocfilters==1.4.2
parsimonious==0.8.0
parso==0.3.1
patsy==0.5.1
pbkdf2==1.3
persistent==4.2.4.2
pexpect==4.6.0
pickleshare==0.7.5
pigar==0.9.0
plotly==4.0.0
pluggy==0.13.0
prometheus-client==0.4.1
prompt-toolkit==2.0.5
protobuf==3.6.1
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
py-ecc==1.4.3
py-solc==2.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodome==3.7.0
pyethash==0.1.27
pyethereum==1.0.0
Pygments==2.2.0
pyparsing==2.2.0
pysha3==1.0.2
pytest==5.2.0
python-dateutil==2.7.5
python-engineio==1.7.0
python-hmac-auth==0.4
python-socketio==1.8.1
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
ratelimit==2.2.1
repoze.lru==0.7
requests==2.21.0
requests-http-signature==0.1.0
requests-oauthlib==0.8.0
retrying==1.3.3
rlp==1.0.3
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.2
scipy==1.2.0
scrypt==0.8.6
semantic-version==2.6.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
statsmodels==0.10.1
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
transaction==2.2.0
translationstring==1.3
tweepy==3.5.0
urllib3==1.22
wcwidth==0.1.7
web3==4.8.2
webencodings==0.5.1
websockets==6.0
Werkzeug==0.12.2
widgetsnbextension==3.4.2
xlrd==1.2.0
zc.lockfile==1.2.1
ZConfig==3.2.0
zipp==0.6.0
ZODB==5.3.0
zodbpickle==1.0
zope.interface==4.4.3dash==1.6.1
dash-core-components==1.5.1
dash-html-components==1.0.2
dash-renderer==1.2.1
dash-table==4.5.1
